requirements.txt
amqp==2.2.1
anyjson==0.3.3
asn1crypto==0.22.0
bcrypt==3.1.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
bs4==0.0.1
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
contextlib2==0.5.5
cryptography==1.9
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-address==0.1.5
django-carton==1.2.1
django-ckeditor==5.2.2
django-custom-user==0.7
django-debug-toolbar==1.8
django-flat-responsive==1.4.1
django-health-check==2.3.0
django-js-asset==0.1.1
django-paypal==0.3.6
ecdsa==0.13
Fabric3==1.13.1.post1
future==0.16.0
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
oauthlib==2.0.2
olefile==0.44
paramiko==2.2.1
Pillow==4.2.1
psycopg2==2.7.3
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pycryptodomex==3.4.6
pyjwkest==1.3.2
PyJWT==1.5.0
PyNaCl==1.1.2
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
raven==6.1.0
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.5.0
simple-bank-korea==0.2.3
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
sqlparse==0.2.3
urllib3==1.22
vine==1.1.4
whitenoise==3.3.0
